requirement.txt
Django==3.1.4
djangorestframework==3.11.0
requests